+++ Shared/Views/SimulationView/Graphic Menu Views/Balance Sheet/BS Global ChartView.swift
@@ -181,7 +181,7 @@ struct BalanceSheetLineChartView: NSUIViewRepresentable {
         // ajouter les DataSet au Chartdata
         let data = LineChartData(dataSets: dataSets)
         data.setValueTextColor(ChartThemes.DarkChartColors.valueColor)
-        data.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: CGFloat(12.0))!)
+        data.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: 12.0)!)
         data.setValueFormatter(DefaultValueFormatter(formatter: valueKiloFormatter))
 
         // ajouter le Chartdata au ChartView
