+++ Shared/Views/SimulationView/Graphic Menu Views/Impots Views/IrppEvolutionChartView.swift
@@ -139,7 +139,7 @@ struct IrppEvolutionLineChartView: UIViewRepresentable {
         // ajouter les DataSet au ChartData
         let lineChartData = LineChartData(dataSets: lineDataSets)
         lineChartData.setValueTextColor(ChartThemes.DarkChartColors.valueColor)
-        lineChartData.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: CGFloat(12.0))!)
+        lineChartData.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: 12.0)!)
         lineChartData.setValueFormatter(DefaultValueFormatter(formatter: decimalX100IntegerFormatter))
 
         // BARRES
@@ -148,7 +148,7 @@ struct IrppEvolutionLineChartView: UIViewRepresentable {
         // ajouter les DataSet au ChartData
         let barChartData = BarChartData(dataSets: barDataSets)
         barChartData.setValueTextColor(ChartThemes.DarkChartColors.valueColor)
-        barChartData.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: CGFloat(12.0))!)
+        barChartData.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: 12.0)!)
         barChartData.setValueFormatter(DefaultValueFormatter(formatter: decimalFormatter))
 
         // combiner les ChartData
@@ -258,7 +258,7 @@ struct IrppTranchesLineChartView: UIViewRepresentable {
         // ajouter les DataSet au Chartdata
         let data = LineChartData(dataSets: dataSets)
         data.setValueTextColor(ChartThemes.DarkChartColors.valueColor)
-        data.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: CGFloat(12.0))!)
+        data.setValueFont(NSUIFont(name: "HelveticaNeue-Light", size: 12.0)!)
         data.setValueFormatter(DefaultValueFormatter(formatter: valueKiloFormatter))
 
         // ajouter le Chartdata au ChartView
